backend/core/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-graphql-jwt==0.4.0
django-simple-history==3.7.0
django-storages==1.14.4
djangorestframework==3.15.2
Django==5.1.1
google-api-core==2.21.0
google-cloud-logging==3.11.3
google-cloud-storage==2.18.2
google-cloud-tasks==2.16.5
graphene==3.3
graphene-django==3.2.2
graphene-django-optimizer==0.10.0
graphql-core==3.2.4
graphql-relay==3.2.0
hashids==1.3.1
inflection==0.5.1
packaging==24.1
promise==2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.12.4
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.14.1
uvicorn==0.32.0
whitenoise==6.7.0
workos==5.32.0
